requirements.txt
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.4.0
astroid==3.2.2
bcrypt==4.1.3
beanie==1.26.0
black==24.4.2
certifi==2024.6.2
cffi==1.16.0
click==8.1.7
cryptography==42.0.8
dill==0.3.8
distro==1.9.0
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.1.2
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.4
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
isort==5.13.2
Jinja2==3.1.4
lazy-model==0.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mccabe==0.7.0
mdurl==0.1.2
motor==3.4.0
mypy-extensions==1.0.0
openai==1.34.0
orjson==3.10.5
packaging==24.1
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.2.2
pyasn1==0.6.0
pycodestyle==2.12.0
pycparser==2.22
pydantic==2.7.4
pydantic-settings==2.3.3
pydantic_core==2.18.4
Pygments==2.18.0
pylint==3.2.3
pymongo==4.7.3
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.1
rich==13.7.1
rsa==4.9
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.5
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.12.2
ujson==5.10.0
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
